modif type parametre compteur symbole en unint32_t

+++ ProjetL3S1/relocalisation/write_file.c
@@ -82,7 +82,7 @@ void write_section_header (FILE *f, Elf32_Shdr *section_header_table, Elf32_Ehdr
 
 }
 
-void write_symbole_table (FILE *f, Elf32_Sym *symb_table,Elf32_Ehdr header,Elf32_Shdr *section_headers_table, uint16_t symbols_count ){
+void write_symbole_table (FILE *f, Elf32_Sym *symb_table,Elf32_Ehdr header,Elf32_Shdr *section_headers_table, uint32_t symbols_count ){
 
     Elf32_Half symtable_index = 0;
     int i = 0;
